requirements.txt
streamlit
torch
torchvision
numpy
pandas
plotly
Pillow
